+++ R/args.R
@@ -1,32 +1,46 @@
 process_args <- function(args, app) {
   app <- as_app(app)
+  app_opts <- app$opts
+  app_args <- app$args
+  app_commands <- app$commands
+  app_meta <- if (length(app$data)) {
+    prune_empty(as.list(unclass(app$data)))
+  }
+  help_scope <- list(list(
+    name = app$data$name %||% basename(app$filepath),
+    opts = app$opts,
+    args = app$args,
+    commands = app$commands %||% list(),
+    meta = app_meta
+  ))
 
   if (!inherits(args, "connection")) {
     args <- textConnection(args)
     on.exit(close(args))
   }
 
-  short_opt_to_long_opt <- local({
-    table <- unlist(lapply(app$opts, function(opt) opt$short))
-    table <- setNames(
-      as.list(sprintf("--%s", names(table))),
-      sprintf("-%s", table)
-    )
-    function(short_opt) table[[short_opt]]
-  })
+  short_opt_to_long_opt <- function(short_opt) {
+    short <- str_drop_prefix(short_opt, "-")
+    for (i in seq_along(app_opts)) {
+      if (identical(short, app_opts[[i]]$short)) {
+        return(paste0("--", names(app_opts)[[i]]))
+      }
+    }
+  }
+
   positional_args <- character()
   while (length(a <- readLines(args, 1L))) {
     if (a == "--" || a == "--args") {
       break
     }
 
-    if (a == "--help") {
-      print_app_help(app, yaml = "--yaml" %in% readLines(args))
-      if (interactive()) {
-        return(invisible(FALSE))
-      } else {
-        q("no")
-      }
+    if (a %in% c("--help", "--help-yaml")) {
+      print_app_help(
+        app,
+        scope = help_scope,
+        yaml = a == "--help-yaml" || "--yaml" %in% readLines(args)
+      )
+      return(if (interactive()) invisible(FALSE) else q("no"))
     }
 
     arg_type <-
@@ -34,18 +48,29 @@ process_args <- function(args, app) {
         "long-opt"
       } else if (startsWith(a, "-")) {
         "short-opt"
+      } else if (a %in% names(app_commands)) {
+        "command"
       } else {
-        # if (a %in% names(app$commands)) "command" else
         "positional-arg"
       }
 
     if (arg_type == "command") {
-      .NotYetImplemented()
-      # if(length(app$args)) {
-      #   pushBack(app$args, args)
-      #   app$args <- NULL
-      # }
-      return(process_args(args, app = app$commands[[a]]))
+      app$exprs[[app_commands$.val_pos_in_exprs]] <- a
+      command <- app_commands[[a]]
+      append(app_opts) <- command$opts
+      append(app_args) <- command$args
+      command_meta <- if (!is.null(command$meta)) {
+        prune_empty(as.list(unclass(command$meta)))
+      }
+      help_scope[[length(help_scope) + 1L]] <- list(
+        name = a,
+        opts = command$opts,
+        args = command$args,
+        commands = command$commands %||% list(),
+        meta = command_meta
+      )
+      app_commands <- command$commands
+      next
     }
 
     if (arg_type == "positional-arg") {
@@ -70,18 +95,18 @@ process_args <- function(args, app) {
       name <- substring(a, 3, equals_idx - 1L)
       name <- gsub("-", "_", name, fixed = TRUE)
       val <- str_drop_prefix(a, equals_idx)
-      spec <- app$opts[[name]]
+      spec <- app_opts[[name]]
     } else {
       # --name
       name <- str_drop_prefix(a, "--")
       name <- gsub("-", "_", name, fixed = TRUE)
 
-      spec <- app$opts[[name]]
+      spec <- app_opts[[name]]
 
       # if flag not known, maybe this is a switch flag
       if (is.null(spec) && startsWith(a, "--no-")) {
         alt_name <- str_drop_prefix(name, "no_")
-        spec <- app$opts[[alt_name]]
+        spec <- app_opts[[alt_name]]
         if (!is.null(spec)) {
           val <- "false"
           name <- alt_name
@@ -145,7 +170,7 @@ process_args <- function(args, app) {
   if (length(positional_args)) {
     # we've parsed all the command line args,
     # we can now match positional args
-    specs <- app$args
+    specs <- app_args %||% structure(list(), names = character())
 
     collector <- which(
       endsWith(names(specs), "...") |
@@ -171,7 +196,7 @@ process_args <- function(args, app) {
       }
     }
 
-    if (length(specs) != length(positional_args)) {
+    if (length(specs) < length(positional_args)) {
       stop(
         "Arguments not recognized: ",
         paste0(positional_args[-seq_along(specs)], collapse = " ")
@@ -189,117 +214,4 @@ process_args <- function(args, app) {
   }
 
   invisible(TRUE)
-}
-
-
-print_app_help <- function(app, yaml = TRUE) {
-  app <- as_app(app)
-  if (yaml) {
-    x <- c(app$data, list(options = app$opts), list(arguments = app$args))
-    for (nm in names(x$options)) {
-      x$options[[nm]]$.val_pos_in_exprs <- NULL
-    }
-
-    for (nm in names(x$arguments)) {
-      x$arguments[[nm]]$.val_pos_in_exprs <- NULL
-    }
-
-    print.yaml(x)
-    return()
-  }
-
-  app_name <- app$data$name %||% basename(app$filepath)
-  usage <- paste0(
-    collapse = " ",
-    c(
-      "Usage:",
-      app_name,
-      if (length(app$opts)) "[options]",
-      if (length(app$args)) {
-        local({
-          # "<arguments>"
-          x <- paste0(sprintf("<%s>", names(app$args)), collapse = " ")
-          x <- sub("<...", "...<", x, fixed = TRUE)
-          x <- sub("...>", ">...", x, fixed = TRUE)
-          x <- paste0("[", x, "]")
-          x
-        })
-      }
-    )
-  )
-  description <- app$data$description
-
-  options <- imap_chr(app$opts, function(opt, name) {
-    if (opt$arg_type == "switch") {
-      true <- paste0("--", name)
-      false <- paste0("--no-", name)
-      default <- if (opt$default) true else false
-      header <- sprintf("  %s | %s  (Default: %s)", true, false, default)
-      description <- if (length(opt$description)) {
-        strwrap(opt$description, 70, indent = 6, exdent = 6)
-      }
-      out <- paste0(c(header, description), collapse = "\n")
-      return(out)
-    }
-
-    if (opt$arg_type == "option") {
-      default <- opt$default
-      type <- opt$val_type
-      short_name <- opt$short
-      if (type == "string") {
-        default <- deparse1(default)
-      }
-      header <- if (is.null(short_name)) {
-        sprintf("  --%s <value>  (Default: %s, Type: %s)", name, default, type)
-      } else {
-        sprintf(
-          "  -%s, --%s <value>  (Default: %s, Type: %s)",
-          short_name,
-          name,
-          default,
-          type
-        )
-      }
-      description <- if (length(opt$description)) {
-        strwrap(opt$description, 70, indent = 6, exdent = 6)
-      }
-      out <- paste0(c(header, description), collapse = "\n")
-      return(out)
-    }
-
-    # noisy safe fallback (shouldn't happen)
-    opt$.val_pos_in_exprs <- NULL
-    yaml::as.yaml(opt)
-  })
-
-  ## omit positional args without a description (they appear in usage)
-  args <- do.call(
-    c,
-    imap(app$args, function(arg, name) {
-      if (!length(arg$description)) {
-        return()
-      }
-
-      description <- if (length(arg$description)) {
-        strwrap(arg$description, 70, indent = 6, exdent = 6)
-      }
-      header <- paste0("  ", sub("^\\.\\.\\.|\\.\\.\\.$", "", name))
-      paste0(c(header, description), collapse = "\n")
-    })
-  )
-
-  if (length(options)) {
-    options <- c("\nOptions:", options)
-  }
-
-  if (length(args)) {
-    args <- c("\nArguments:", args)
-  }
-
-  if (length(description)) {
-    description <- c(description, "")
-  }
-
-  writeLines(c(description, usage, options, args))
-  return()
 }